requirements.txt
fastapi
uvicorn[standard]
httpx
pyyaml
slowapi
python-dotenv
tenacity
redis[asyncio]